aquila-fastapi/requirements.txt
numpy
pandas
anndata
scipy
gpflow
patsy
NaiveDE
python-igraph
pointpats
fastapi[full]
uvicorn
